+++ Langscaper/App.axaml
@@ -1,7 +1,7 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             x:Class="CSP.App"
-             xmlns:local="using:CSP"
+             x:Class="Langscaper.App"
+             xmlns:local="using:Langscaper"
              RequestedThemeVariant="Default">
              <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
