+++ Langscaper/Views/MainWindow.axaml
@@ -1,10 +1,10 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="using:CSP.ViewModels"
+        xmlns:vm="using:Langscaper.ViewModels"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="CSP.Views.MainWindow"
+        x:Class="Langscaper.Views.MainWindow"
         x:DataType="vm:MainWindowViewModel"
         Icon="/Assets/Langscaper.png"
         Title="Langscaper">
